refactoring: rename class: AbstractDataFile -> DataFile
