+++ lib/redshift_connector/data_file.rb
@@ -1,7 +1,32 @@
-# ensure to create module
-module RedshiftConnector
-end
+require 'redshift_connector/logger'
+require 'zlib'
 
-require "redshift_connector/reader"
-require "redshift_connector/url_data_file_bundle"
-require "redshift_connector/logger"
+module RedshiftConnector
+  class DataFile
+    def initialize(reader_class:)
+      @reader_class = reader_class
+    end
+
+    def each_row(&block)
+      f = open
+      begin
+        if gzipped_object?
+          f = Zlib::GzipReader.new(f)
+        end
+        @reader_class.new(f).each(&block)
+      ensure
+        f.close
+      end
+    end
+
+    # abstract open
+
+    def data_object?
+      @reader_class.data_object?(key)
+    end
+
+    def gzipped_object?
+      File.extname(key) == '.gz'
+    end
+  end
+end
